게시판 조회, 전체조회 추가 및 전체조회 테스트 추가

+++ backend/src/main/java/com/studygram/domain/Post.java
@@ -7,7 +7,7 @@ import java.util.Date;
 @Getter
 @Setter
 public class Post {
-    long idx;
+    int idx;
     int userId;
     int imageUrlId;
     String content;
